Remove job source from local blobstores after applying is successful
This change updates the bosh-agent to delete blobs associated with a
job's rendered templates after successfully applying and installing the
job. Previously, we never cleaned up these blobs and the growth of both
`/var/vcap/data/sensistive_blobs` and `/var/vcap/data/blobs` was
unbounded.

This change also updates the cascading blobstore to not attempt deletion
from the external blobstore. The external blobstore does not support
deletion, and the agent should not be responsible for managing blobs in
the director's external blobstore.

[finishes #162434271](https://www.pivotaltracker.com/story/show/162434271)

+++ agent/blobstore/cascading_blobstore.go
@@ -56,17 +56,15 @@ func (b cascadingBlobstore) Validate() error {
 }
 
 func (b cascadingBlobstore) Delete(blobID string) error {
-	var err error
 	for _, blobManager := range b.blobManagers {
-		err = blobManager.Delete(blobID)
-		if err == nil {
-			break
+		if blobManager.BlobExists(blobID) {
+			return blobManager.Delete(blobID)
 		}
 	}
 
-	if err != nil {
-		return err
-	}
-
-	return b.innerBlobstore.Delete(blobID)
+	// We do not delete from the inner blobstore as external blobstores do not
+	// currently support deletion. Also, the bosh-agent should not be responsible
+	// for deleting blobs from the external blobstore which is managed by the
+	// BOSH Director.
+	return nil
 }
